Refactored and moved TextChannel purge function
Refactored and moved the TextChannel purge function into the
discord.abc.Messageable class so that it can be used in DMChannel
instances and any other object representing a messageable location.
